fix: package reference causes compilation failure
